+++ R/mirtCAT.R
@@ -261,6 +261,9 @@
 #'      
 #'    \item{\code{stem_dims}}{numeric vector of length 2 corresponding to image stem width 
 #'      (in pixels). Default is \code{c(1000, 1000)} for the width and height}
+#'      
+#'    \item{forced_choice}{logical; require a respones to each item? Default is \code{TRUE}.
+#'      This should only be set to \code{FALSE} for surveys (not CATs)}
 #'   
 #' }
 #'